backend/mohamed/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
Markdown==3.6
packaging==24.0
pillow==11.0.0
PyJWT==2.8.0
pysslcmz==1.1.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
sqlparse==0.5.0
sslcommerz-lib==1.0
uritemplate==4.1.1
urllib3==2.2.1
setuptools==80.9.0